+++ packages/core/core/src/factories/useUserFactory.ts
@@ -1,6 +1,6 @@
 import { Ref, computed } from '@vue/composition-api';
 import { UseUser } from '../types';
-import { sharedRef } from '../utils';
+import { sharedRef, Logger, mask } from '../utils';
 
 export interface UseUserFactoryParams<USER, UPDATE_USER_PARAMS, REGISTER_USER_PARAMS> {
   loadUser: () => Promise<USER>;
@@ -22,6 +22,7 @@ export const useUserFactory = <USER, UPDATE_USER_PARAMS, REGISTER_USER_PARAMS ex
 
   const setUser = (newUser: USER) => {
     sharedRef('useUser-user').value = newUser;
+    Logger.debug('useUserFactory.setUser', newUser);
   };
 
   const useUser = (): UseUser<USER, UPDATE_USER_PARAMS> => {
@@ -30,10 +31,14 @@ export const useUserFactory = <USER, UPDATE_USER_PARAMS, REGISTER_USER_PARAMS ex
     const isAuthenticated = computed(() => Boolean(user.value));
 
     const updateUser = async (params: UPDATE_USER_PARAMS) => {
+      Logger.debug('useUserFactory.updateUser', params);
+
       loading.value = true;
       try {
         user.value = await factoryParams.updateUser({currentUser: user.value, updatedUserData: params});
       } catch (err) {
+        Logger.error('useUserFactory.updateUser', err);
+
         throw new Error(err);
       } finally {
         loading.value = false;
@@ -41,10 +46,14 @@ export const useUserFactory = <USER, UPDATE_USER_PARAMS, REGISTER_USER_PARAMS ex
     };
 
     const register = async (registerUserData: REGISTER_USER_PARAMS) => {
+      Logger.debug('useUserFactory.register', registerUserData);
+
       loading.value = true;
       try {
         user.value = await factoryParams.register(registerUserData);
       } catch (err) {
+        Logger.error('useUserFactory.register', err);
+
         throw new Error(err);
       } finally {
         loading.value = false;
@@ -55,10 +64,14 @@ export const useUserFactory = <USER, UPDATE_USER_PARAMS, REGISTER_USER_PARAMS ex
       username: string;
       password: string;
     }) => {
+      Logger.debug('useUserFactory.login', loginUserData);
+
       loading.value = true;
       try {
         user.value = await factoryParams.logIn(loginUserData);
       } catch (err) {
+        Logger.error('useUserFactory.login', err);
+
         throw new Error(err);
       } finally {
         loading.value = false;
@@ -66,19 +79,27 @@ export const useUserFactory = <USER, UPDATE_USER_PARAMS, REGISTER_USER_PARAMS ex
     };
 
     const logout = async () => {
+      Logger.debug('useUserFactory.logout');
+
       try {
         await factoryParams.logOut();
         user.value = null;
       } catch (err) {
+        Logger.error('useUserFactory.err');
+
         throw new Error(err);
       }
     };
 
     const changePassword = async (currentPassword: string, newPassword: string) => {
+      Logger.debug('useUserFactory.changePassword', { currentPassword: mask(currentPassword), newPassword: mask(newPassword) });
+
       loading.value = true;
       try {
         user.value = await factoryParams.changePassword({currentUser: user.value, currentPassword, newPassword});
       } catch (err) {
+        Logger.error('useUserFactory.changePassword', err);
+
         throw new Error(err);
       } finally {
         loading.value = false;
@@ -86,10 +107,14 @@ export const useUserFactory = <USER, UPDATE_USER_PARAMS, REGISTER_USER_PARAMS ex
     };
 
     const load = async () => {
+      Logger.debug('useUserFactory.refreshUser');
       loading.value = true;
+
       try {
         user.value = await factoryParams.loadUser();
       } catch (err) {
+        Logger.error('useUserFactory.refreshUser', err);
+
         throw new Error(err);
       } finally {
         loading.value = false;