+++ packages/core/core/src/types.ts
@@ -443,3 +443,10 @@ export interface AgnosticFacetSearchParams {
   metadata?: any;
   [x: string]: any;
 }
+
+export interface VSFLogger {
+  debug(message?: any, ...args: any): void;
+  info(message?: any, ...args: any): void;
+  warn(message?: any, ...args: any): void;
+  error(message?: any, ...args: any): void;
+}